feat: rewrite Immutable* to drop metaclasses
The only reason this was metaclasses- and the inject form- was to work
around py2k.  The metaclasses are no longer necessary, thus use this
form that just is simple inheritance, and does the nasty as needed.
